Write readme file and reorganize ui test structure
