+++ package.json
@@ -4,11 +4,11 @@
   "description": "Challenge QA Automation",
   "main": "index.js",
   "scripts": {
-    "test:chrome": "npx playwright test --project=chromium",
-    "test:firefox": "npx playwright test --project=firefox",
-    "test:allBrowsers": "npm run test:chrome && npm run test:firefox",
+    "test:ui:chrome": "npx playwright test tests/ui --project=chromium",
+    "test:ui:firefox": "npx playwright test tests/ui --project=firefox",
+    "test:ui:allBrowsers": "npm run test:ui:chrome && npm run test:ui:firefox",
     "test:api": "npx playwright test tests/api --project=ApiTests",
-    "test:all": "npm run test:allBrowsers && npm run test:api"
+    "test:all": "npm run test:ui:allBrowsers && npm run test:api"
   },
   "keywords": [],
   "author": "",
